refac: Rename store
