+++ src/app/[variants]/(main)/resource/(home)/_layout/Body/LibraryList/Item/index.tsx
